Savestate the current "working"
This is a savestate of the current working iteration of the azAD implementation.

Because BOF programming is hard there is some issues we're slowly sussing out with the iterative debugging. Currently the main issue is avoiding the random "no slot for function (reduce number of Win32 APIs called)" that appears when certain funcs are being resolved. I know of two one being `free`. 

Another issue is we're trying to add safe manners to do shit so we need to implement more dynamic memory allocs and their corresponding checks and unhardcode some values and do memsets